backend/requirements.txt
livekit-agents[openai,silero,deepgram,turn-detector]~=1.0
python-dotenv==1.0.0
flask[async]
flask==3.0.0
flask-cors==4.0.0
uvicorn